// apps/web/src/api/__generated__/types/GetProjectResponseRole.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */

/**
 * 현재 사용자 역할
 * @nullable
 */
export type GetProjectResponseRole =
  | (typeof GetProjectResponseRole)[keyof typeof GetProjectResponseRole]
  | null;

export const GetProjectResponseRole = {
  Admin: 'ADMIN',
  Member: 'MEMBER',
} as const;