// apps/web/src/api/__generated__/types/GetProjectResponseSummaryJoinPolicy.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */

/**
 * 프로젝트 참여 정책
 */
export type GetProjectResponseSummaryJoinPolicy =
  (typeof GetProjectResponseSummaryJoinPolicy)[keyof typeof GetProjectResponseSummaryJoinPolicy];

export const GetProjectResponseSummaryJoinPolicy = {
  Open: 'OPEN',
  Request: 'REQUEST',
  Invite: 'INVITE',
} as const;